- update XML parsing for CSC'09
git-svn-id: svn://svn.code.sf.net/p/choco/code/trunk@1077 5779b9a9-2447-0410-887e-bb98d57c6de4

+++ choco-cp/src/main/java/parser/absconparseur/components/PConstraint.java
@@ -73,7 +73,7 @@ public abstract class PConstraint {
 	 * For CSP, returns 0 is the constraint is satified and 1 if the constraint is violated. <br>
 	 * For WCSP, returns the cost for the given tuple.
 	 */
-	public abstract int computeCostOf(int[] tuple);
+	public abstract long computeCostOf(int[] tuple);
 
 	public String toString() {
 		String s = "  constraint " + name + " with arity = " + scope.length + ", scope = ";
